src/requirements.txt
pydantic==2.6.1
discord.py==2.3.2
motor==3.3.2